+++ src/Sentry/SentryOptionsExtensions.cs
@@ -55,8 +55,16 @@ namespace Sentry
         /// Disables the capture of errors through <see cref="TaskScheduler.UnobservedTaskException"/>.
         /// </summary>
         /// <param name="options">The SentryOptions to remove the integration from.</param>
+        [Obsolete("Method has been renamed to DisableUnobservedTaskExceptionCapture.  Please update usage.")]
         public static void DisableTaskUnobservedTaskExceptionCapture(this SentryOptions options) =>
-            options.RemoveIntegration<TaskUnobservedTaskExceptionIntegration>();
+            options.DisableUnobservedTaskExceptionCapture();
+
+        /// <summary>
+        /// Disables the capture of errors through <see cref="TaskScheduler.UnobservedTaskException"/>.
+        /// </summary>
+        /// <param name="options">The SentryOptions to remove the integration from.</param>
+        public static void DisableUnobservedTaskExceptionCapture(this SentryOptions options) =>
+            options.RemoveIntegration<UnobservedTaskExceptionIntegration>();
 
 #if NETFRAMEWORK
         /// <summary>